+++ include/distortos/DynamicThread.hpp
@@ -2,7 +2,7 @@
  * \file
  * \brief DynamicThread class header
  *
- * \author Copyright (C) 2015-2016 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ * \author Copyright (C) 2015-2017 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
  *
  * \par License
  * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
@@ -110,9 +110,10 @@ public:
 	 * \param [in] signalNumber is the signal that will be generated, [0; 31]
 	 *
 	 * \return 0 on success, error code otherwise:
-	 * - EINVAL - \a signalNumber value is invalid;
-	 * - ENOTSUP - reception of signals is disabled for this thread;
 	 * - EINVAL - internal thread object was detached;
+	 * - EINVAL - \a signalNumber value is invalid;
+	 * - ENOMEM - amount of free stack is too small to request delivery of signals;
+	 * - ENOTSUP - reception of signals is disabled for this thread;
 	 *
 	 * \ingroup signals
 	 */
@@ -193,9 +194,10 @@ public:
 	 * \return 0 on success, error code otherwise:
 	 * - EAGAIN - no resources are available to queue the signal, maximal number of signals is already queued in
 	 * associated queue of SignalInformation objects;
-	 * - EINVAL - \a signalNumber value is invalid;
-	 * - ENOTSUP - reception or queuing of signals are disabled for this thread;
 	 * - EINVAL - internal thread object was detached;
+	 * - EINVAL - \a signalNumber value is invalid;
+	 * - ENOMEM - amount of free stack is too small to request delivery of signals;
+	 * - ENOTSUP - reception or queuing of signals are disabled for this thread;
 	 *
 	 * \ingroup signals
 	 */