+++ include/distortos/ThisThread-Signals.hpp
@@ -2,7 +2,7 @@
  * \file
  * \brief ThisThread::Signals namespace header
  *
- * \author Copyright (C) 2015-2016 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ * \author Copyright (C) 2015-2017 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
  *
  * \par License
  * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
@@ -47,6 +47,7 @@ namespace Signals
  *
  * \return 0 on success, error code otherwise:
  * - EINVAL - \a signalNumber value is invalid;
+ * - ENOMEM - amount of free stack is too small to request delivery of signals;
  * - ENOTSUP - reception of signals is disabled for current thread;
  */
 
@@ -112,6 +113,7 @@ SignalSet getSignalMask();
  * - EAGAIN - no resources are available to queue the signal, maximal number of signals is already queued in
  * associated queue of SignalInformation objects;
  * - EINVAL - \a signalNumber value is invalid;
+ * - ENOMEM - amount of free stack is too small to request delivery of signals;
  * - ENOTSUP - reception or queuing of signals are disabled for current thread;
  */
 
@@ -148,6 +150,7 @@ std::pair<int, SignalAction> setSignalAction(uint8_t signalNumber, const SignalA
  * \param [in] signalMask is the SignalSet with new signal mask for current thread
  *
  * \return 0 on success, error code otherwise:
+ * - ENOMEM - amount of free stack is too small to request delivery of signals;
  * - ENOTSUP - reception or catching/handling of signals are disabled for current thread;
  */