+++ include/distortos/internal/synchronization/SignalsCatcherControlBlock.hpp
@@ -2,7 +2,7 @@
  * \file
  * \brief SignalsCatcherControlBlock class header
  *
- * \author Copyright (C) 2015 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ * \author Copyright (C) 2015-2017 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
  *
  * \par License
  * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
@@ -100,6 +100,7 @@ public:
 	 *
 	 * \return 0 on success, error code otherwise:
 	 * - EINVAL - \a signalNumber value is invalid;
+	 * - ENOMEM - amount of free stack is too small to request delivery of signals;
 	 */
 
 	int postGenerate(uint8_t signalNumber, ThreadControlBlock& threadControlBlock);
@@ -129,9 +130,12 @@ public:
 	 * \param [in] signalMask is the SignalSet with new signal mask for associated thread
 	 * \param [in] owner selects whether delivery of signals will be requested if any pending signal is unblocked
 	 * (pointer to owner SignalsReceiverControlBlock object) or not (nullptr)
+	 *
+	 * \return 0 on success, error code otherwise:
+	 * - ENOMEM - amount of free stack is too small to request delivery of signals;
 	 */
 
-	void setSignalMask(SignalSet signalMask, const SignalsReceiverControlBlock* owner);
+	int setSignalMask(SignalSet signalMask, const SignalsReceiverControlBlock* owner);
 
 	SignalsCatcherControlBlock(const SignalsCatcherControlBlock&) = delete;
 	SignalsCatcherControlBlock(SignalsCatcherControlBlock&&) = default;
@@ -179,9 +183,12 @@ private:
 	 * already pending. The thread is unblocked if it was blocked.
 	 *
 	 * \param [in] threadControlBlock is a reference to associated ThreadControlBlock
+	 *
+	 * \return 0 on success, error code otherwise:
+	 * - ENOMEM - amount of free stack is too small to request delivery of signals;
 	 */
 
-	void requestDeliveryOfSignals(ThreadControlBlock& threadControlBlock);
+	int requestDeliveryOfSignals(ThreadControlBlock& threadControlBlock);
 
 	/// storage for Association objects
 	StorageUniquePointer storageUniquePointer_;