+++ CHANGELOG.md
@@ -44,6 +44,9 @@ is very likely to be used in the application (for example to zero-initialize var
 - Reduced default size of stack for interrupts to 1kB.
 - Reduced size of stack for interrupts to 1kB in all test configurations.
 - Reduced size of stack for idle thread to 256 bytes when support for thread detachment is enabled.
+- `architecture::requestFunctionExecution()` checks for amount of free stack before doing any stack modifications. If
+there's not enough free stack available, it returns an `ENOMEM` error code. Modify all callers of this function (which
+includes functions to generate/queue signals and set signal mask) to handle this error code.
 
 ### Fixed